This will be the replace bad words functions, getting seg fault when I run it
I had it working and then seg faults started, when it did work though it wouldn't update the files. I put statements in to make sure it passed correct information and to make sure it was entering the correct loops but somewhere in the replacing process it failed. Need a fresh pair of eyes on this (hint hint Jordan genius skills) to see if yall can find the error.